serialize props
